# Breakdown generic writeOffsetTable