This is a broken down piece of https://github.com/prometheus/tsdb/pull/627

With this, I am avoiding allocating slice headers for postings by having the name and value as separate fields. This saves quite some allocs.

[Don't merge until Prometheus 2.11 is out]