@cstyan 2019-07-02
> This saves quite some allocs.

Any sense of how much? 10%? 20%?