@codesome 2019-07-03
I don't have numbers for individual optimizations. All combined numbers is in this comment https://github.com/prometheus/tsdb/pull/627#issuecomment-503453181. I will try to get number for each optimizations.